+++ backend/src/main/java/com/jamddo/lotto/repository/WinInfoRepositoryCustomImpl.java
@@ -57,4 +57,5 @@ public class WinInfoRepositoryCustomImpl implements WinInfoRepositoryCustom{
                 .orderBy(winInfo.id.desc())
                 .fetchFirst();
     }
+
 }
